# mes-root/docs/mes-auto_swagger.yaml
openapi: 3.0.0
...
servers:
  - description: DEV
    url: https://localhost:9998
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/hengyi/mes-auto/1.0.0
info:
  description: auto api
  version: "1.0.0"
  title: Mes Auto
  contact:
    email: ixtf1984@gmail.com
...
paths:
...
  /api/temporaryBoxs:
    post:
      tags:
        - admin
      summary: 新建暂存箱
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporaryBoxUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryBox'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      summary: 暂存箱列表
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemporaryBox'
        '400':
          $ref: '#/components/responses/JException'
...
  /api/temporaryBoxs/{id}:
    put:
      tags:
        - admin
      summary: 更新暂存箱
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporaryBoxUpdateCommand'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryBox'
        '400':
          $ref: '#/components/responses/JException'
    get:
      tags:
        - config
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryBox'
        '400':
          $ref: '#/components/responses/JException'
...
components:
  securitySchemes:
    # arbitrary name for the security scheme
    bearerAuth:
      type: http
      scheme: bearer
      # optional, arbitrary value for documentation purposes
      bearerFormat: JWT
  responses:
...
    JException:
      description: |
        非200的异常一般都会处理过，没处理的都是http status默认

        如：401没认证，403没权限，404找不到，等等
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JException'
  schemas:
    JException:
      type: object
      properties:
        errorCode:
          type: string
          example: E00000
        errorMessage:
          type: string
...
    EntityDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: string
...
    Corporation:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    Workshop:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        note:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
...
    Product:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
...
    Batch:
      type: object
      description: 批号
      properties:
        id:
          type: string
        batchNo:
          type: string
        silkWeight:
          type: number
        centralValue:
          type: number
        holeNum:
          type: number
        spec:
          type: string
          description: 规格
        tubeColor:
          type: string
          description: 规格
        note:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
...
    Grade:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sortBy:
          type: string
          description: 分数，丝锭的等级需要100分，才能推入自动包装线
...
    TemporaryBoxUpdateCommand:
      type: object
      required:
        - code
        - batch
        - grade
      properties:
        code:
          type: string
        batch:
          $ref: '#/components/schemas/EntityDTO'
        grade:
          $ref: '#/components/schemas/EntityDTO'
    TemporaryBox:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        batch:
          $ref: '#/components/schemas/Batch'
        grade:
          $ref: '#/components/schemas/Grade'
        silkCount:
          type: integer
          description: 暂存数量
...
security:
  - bearerAuth: []